requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
Django==4.2
django-jazzmin==3.0.0
django-phonenumber-field==8.0.0
frozenlist==1.4.1
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
phonenumberslite==8.13.45
pydantic==2.9.2
pydantic_core==2.23.4
python-decouple==3.8
sqlparse==0.5.1
typing_extensions==4.12.2
yarl==1.12.1
